refactor some SMB methods
