+++ keywords/ssh/common.py
@@ -12,7 +12,7 @@ class Common_SSH:
                       user: str = private_config['SSH_USERNAME'],
                       password: str = private_config['PASSWORD']) -> None:
         """
-        This method adds the files used for testing smb permissions
+        This method adds files to be used for testing
 
         :param file: is the name of the file to add
         :param path: is the path of the file
@@ -25,20 +25,6 @@ class Common_SSH:
             - Common.add_test_file('myFile.txt', 'tank/path', '10.0.0.1', 'user', 'password')
         """
         SSH_Command_Line(f'sudo touch /mnt/{path}/{file}', ip, user, password)
-
-    @classmethod
-    def add_smb_test_files(cls, user: str, dataset_path: str, ip: str) -> None:
-        """
-        This method adds the files used for testing smb permissions
-
-        :param user: is the user to be sending files to the smb share
-        :param dataset_path: is the path if the smb share
-        :param ip: the IP of the smb share box
-        """
-        SSH_Command_Line(f'cd ~; touch putfile', private_config['SMB_ACL_IP'], user, 'testing')
-        SSH_Command_Line(f'cd /mnt/{dataset_path}/; touch getfile', ip, user, 'testing')
-        SSH_Command_Line(f'cd /mnt/{dataset_path}/; touch deletefile', ip, user, 'testing')
-        SSH_Command_Line(f'cd /mnt/{dataset_path}/; echo "touch execfile2.txt" >> /mnt/{dataset_path}/execfile.sh', ip, user, 'testing')
 
     @classmethod
     def assert_file_not_exist(cls, user: str, file: str) -> bool:
@@ -58,15 +44,6 @@ class Common_SSH:
         # don't recreate the file if it already exists.
         if path.exists(shared_config['KEYPATH']) is False:
             Local_Command_Line(f"yes | ssh-keygen -t rsa -f {shared_config['KEYPATH']} -q -N ''")
-
-    @classmethod
-    def delete_smb_test_files(cls, user: str) -> None:
-        """
-        This method deletes the files used for testing smb permissions
-
-        :param user: is the user to be sending files to the smb share
-        """
-        SSH_Command_Line(f'rm * | grep file', private_config['SMB_ACL_IP'], user, 'testing')
 
     @classmethod
     def get_checksum_of_file(cls, ip: str, filename: str, user: str = private_config["SSH_USERNAME"], password: str = private_config["SSH_PASSWORD"]) -> str:
